Add preliminary multi-adventure support
There’s much still to be done in this direction but it’s
a start at least. It seems like sometimes CMSy features
are needed, like the site-specific images used for CR
which should be configured for a particular instance
rather than the adventure itself.

The env-setting in tests is thanks to @jtomaszewski:
https://elixirforum.com/t/using-application-get-env-application-put-env-in-exunit-tests/8019/3