ud: Update XML renderer to include all prompts
The aspect where the wrapped prompts have to be serialised
into a string that’s then deserialised when rendering seems
wasteful but I couldn’t figure out another way since a
trait can’t be a struct.

Also this is effectively caching the recorded prompts
which may prove problematic without a way to bust.